+++ CHANGELOG.md
@@ -11,7 +11,10 @@
 
 ### Fixed
 - Boreholes could not be exported in the _view_ environment.
+- `Boreholes.swissgeol.ch ID` was not listed in the location type ID filter.
 - The lithostratigraphy was missing when copying a stratigraphy.
+- Loading a page by url did not work after a redirect from the login page.
+- Editing boreholes with photos would cause an error.
 
 ## v2.1.1200 - 2025-05-22